App: both windows draw the shared state and nothing else
ui = f(state). Both row actions now take a RowActionState and plain closures,
never an AppListModel: the model's init registers notification permission, arms
timers and installs FS watchers, so a view that needs one cannot be rendered by
anything but the running app.

- PopoverRowAction and WorkbenchRowAction are extracted to their own files, with
  AppIconCache and DownloadReadout moved alongside so neither view drags a window
  in with it.
- AppListModel.rowState(for:) assembles the facts; rowRoute(for:) resolves how an
  available update would be applied. The route is decided once, on the model, so
  the two windows cannot disagree about whether a row is one-click at all.
- Everything a view needs to draw is in the state or an argument: the staged
  package's file name, managedHere, the failure message and whether it was the
  rate limit. The popover kept reading its own copies of the first three after
  the ladder moved to Core, which made three pairs of states draw identically.

Two presentation differences stay, both deliberate: the workbench does not
one-click a major upgrade or a gated App Store row, because the licence-boundary
and region explanations that make those safe to act on live in the popover. It
names the state and points there — which is not the same as rendering nothing.

One behaviour change worth calling out: the workbench used to show the major-
upgrade warning for every major upgrade, including App Store and detection-only
rows the popover never gated that way. It now applies the popover's gate
verbatim, so the two windows agree. If a non-installable major upgrade should
carry a warning, that belongs in the popover and both windows would follow.

The five new strings reuse keys the catalogue already has rather than adding
near-duplicates of them.
